COSBETA-87: Review comments
