+++ cosmetics-web/app/helpers/cpnp_helper.rb
@@ -394,4 +394,15 @@ module CpnpHelper
       minrangevalue: "minRangeValue",
       maxrangevalue: "maxRangeValue"
   }.freeze
+
+  UNIT_NAME = {
+      less_than_01_percent: "≤ 0.1%",
+      greater_than_01_less_than_1_percent: "> 0.1 - ≤ 1%",
+      greater_than_1_less_than_5_percent: "> 1 - ≤ 5%",
+      greater_than_5_less_than_10_percent: "> 5 - ≤ 10%",
+      greater_than_10_less_than_25_percent: "> 10 - ≤ 25%",
+      greater_than_25_less_than_50_percent: "> 25 - ≤ 50%",
+      greater_than_50_less_than_75_percent: "> 50 - ≤ 75%",
+      greater_than_75_less_than_100_percent: "> 75 - ≤ 100%"
+  }.freeze
 end
