+++ cosmetics-web/app/models/concerns/cpnp_notification_categories.rb
@@ -1,0 +1,116 @@
+module CpnpNotificationCategories
+  def get_category(id)
+    CATEGORY_ID[id]
+  end
+
+  CATEGORY_ID = {
+      100003 => :skin_products,
+      100004 => :skin_care_products,
+      100005 => :face_care_products_other_than_face_mask,
+      100018 => :face_mask,
+      100025 => :eye_contour_products,
+      100032 => :lip_care_products,
+      100035 => :hand_care_products,
+      100044 => :foot_care_products,
+      100049 => :body_care_products,
+      100062 => :external_intimate_care_products,
+      100064 => :chemical_exfoliation_products,
+      100067 => :mechanical_exfoliation_products,
+      100071 => :skin_lightening_products,
+      100079 => :other_skin_care_products,
+      100087 => :skin_cleansing_products,
+      100088 => :soap_products,
+      100092 => :bath_shower_products,
+      100101 => :makeup_remover_products,
+      100105 => :external_intimate_hygiene_products,
+      100108 => :other_skin_cleansing_products,
+      100116 => :body_hair_removal_products,
+      100117 => :chemical_depilatories,
+      100120 => :physical_epilation_products,
+      100124 => :other_body_hair_removal_products,
+      100126 => :bleach_for_body_hair_products,
+      100127 => :bleach_for_body_hair,
+      100130 => :correction_of_body_odour_andor_perspiration,
+      100131 => :products_with_antiperspirant_activity,
+      100138 => :products_without_antiperspirant_activity,
+      100144 => :shaving_and_pre_after_shaving_products,
+      100145 => :shaving_products,
+      100153 => :pre_aftershaving_products,
+      100157 => :other_shaving_and_pre_after_shaving_products,
+      100159 => :makeup_products,
+      100160 => :foundation,
+      100165 => :concealer,
+      100171 => :other_face_makeup_products,
+      100173 => :mascara,
+      100178 => :eye_shadow,
+      100183 => :eye_pencil,
+      100186 => :eye_liner,
+      100192 => :other_eye_makeup_products,
+      100196 => :lip_stick,
+      100199 => :lipstick_sealer,
+      100202 => :other_lip_makeup_products,
+      100210 => :body_or_face_paint_including_carneval_makeup,
+      100213 => :other_makeup_products,
+      100221 => :perfumes,
+      100222 => :hydroalcoholic_perfumes,
+      100225 => :non_hydroalcoholic_perfumes,
+      100230 => :sun_and_selftanning_products,
+      100231 => :before_and_after_sun_products,
+      100239 => :sun_protection_products,
+      100246 => :selftanning_products,
+      100250 => :other_sun_and_selftanning_products,
+      100252 => :other_skin_products,
+      100253 => :other_skin_products_child,
+      100255 => :hair_and_scalp_products,
+      100256 => :hair_and_scalp_care_and_cleansing_products,
+      100257 => :shampoo,
+      100262 => :hair_conditioner,
+      100268 => :scalp_and_hair_roots_care_products,
+      100272 => :antidandruff_products,
+      100278 => :antihairloss_products,
+      100282 => :other_hair_and_scalp_care_and_cleansing_products,
+      100284 => :hair_colouring_products,
+      100285 => :oxidative_hair_colour_products,
+      100293 => :nonoxidative_hair_colour_products,
+      100298 => :hair_bleaching_and_dye_remover_products,
+      100310 => :other_hair_colouring_products,
+      100312 => :hair_styling_products,
+      100313 => :products_for_temporary_hair_styling,
+      100324 => :permanent_wave_products,
+      100332 => :hair_relaxer_straightener_products,
+      100338 => :other_hair_styling_products,
+      100340 => :other_hair_and_scalp_products,
+      100341 => :hair_sun_protection_products,
+      100347 => :other_hair_and_scalp_products_child,
+      100349 => :nail_and_cuticle_products,
+      100350 => :nail_varnish_and_remover_products,
+      100351 => :nail_varnish_nail_makeup,
+      100355 => :nail_varnish_remover,
+      100357 => :nail_varnish_thinner,
+      100359 => :nail_bleach,
+      100361 => :other_nail_varnish_and_remover_products,
+      100363 => :nail_care_nail_hardener_products,
+      100364 => :nail_care_products,
+      100367 => :nail_hardener,
+      100369 => :other_nail_care_nail_hardener_products,
+      100371 => :nail_glue_remover_products,
+      100372 => :nail_glue_remover,
+      100374 => :other_nail_and_cuticle_products,
+      100375 => :cuticle_remover_softener,
+      100377 => :nail_sculpting_products,
+      100379 => :other_nail_and_cuticle_products_child,
+      100381 => :oral_hygiene_products,
+      100382 => :tooth_care_products,
+      100383 => :toothpaste,
+      100386 => :tooth_cleansing_powder_salt,
+      100388 => :other_tooth_care_products,
+      100390 => :mouth_wash_breath_spray,
+      100391 => :mouth_wash,
+      100395 => :breath_spray,
+      100398 => :other_mouth_wash_breath_spray_products,
+      100400 => :tooth_whiteners,
+      100401 => :tooth_whiteners_child,
+      100403 => :other_oral_hygiene_products,
+      100404 => :other_oral_hygiene_products_child
+  }.freeze
+end
