+++ cosmetics-web/app/models/concerns/cpnp_notification_frame_formulations.rb
@@ -1,140 +1,7 @@
-module CpnpNotificationInfo
-  def get_notification_type(id)
-    NOTIFICATION_TYPE_ID[id]
-  end
-
-  def get_category(id)
-    CATEGORY_ID[id]
-  end
-
+module CpnpNotificationFrameFormulations
   def get_frame_formulation(id)
     FRAME_FORMULATION_ID[id]
   end
-
-  def get_trigger_rules_question(id)
-    TRIGGER_RULES_QUESTION_ID[id]
-  end
-
-  def get_trigger_rules_question_element(id)
-    TRIGGER_RULES_QUESTION_ELEMENT_ID[id]
-  end
-
-  NOTIFICATION_TYPE_ID = {
-      1 => :predefined,
-      2 => :exact,
-      3 => :range
-  }.freeze
-
-  CATEGORY_ID = {
-      100003 => :skin_products,
-      100004 => :skin_care_products,
-      100005 => :face_care_products_other_than_face_mask,
-      100018 => :face_mask,
-      100025 => :eye_contour_products,
-      100032 => :lip_care_products,
-      100035 => :hand_care_products,
-      100044 => :foot_care_products,
-      100049 => :body_care_products,
-      100062 => :external_intimate_care_products,
-      100064 => :chemical_exfoliation_products,
-      100067 => :mechanical_exfoliation_products,
-      100071 => :skin_lightening_products,
-      100079 => :other_skin_care_products,
-      100087 => :skin_cleansing_products,
-      100088 => :soap_products,
-      100092 => :bath_shower_products,
-      100101 => :makeup_remover_products,
-      100105 => :external_intimate_hygiene_products,
-      100108 => :other_skin_cleansing_products,
-      100116 => :body_hair_removal_products,
-      100117 => :chemical_depilatories,
-      100120 => :physical_epilation_products,
-      100124 => :other_body_hair_removal_products,
-      100126 => :bleach_for_body_hair_products,
-      100127 => :bleach_for_body_hair,
-      100130 => :correction_of_body_odour_andor_perspiration,
-      100131 => :products_with_antiperspirant_activity,
-      100138 => :products_without_antiperspirant_activity,
-      100144 => :shaving_and_pre_after_shaving_products,
-      100145 => :shaving_products,
-      100153 => :pre_aftershaving_products,
-      100157 => :other_shaving_and_pre_after_shaving_products,
-      100159 => :makeup_products,
-      100160 => :foundation,
-      100165 => :concealer,
-      100171 => :other_face_makeup_products,
-      100173 => :mascara,
-      100178 => :eye_shadow,
-      100183 => :eye_pencil,
-      100186 => :eye_liner,
-      100192 => :other_eye_makeup_products,
-      100196 => :lip_stick,
-      100199 => :lipstick_sealer,
-      100202 => :other_lip_makeup_products,
-      100210 => :body_or_face_paint_including_carneval_makeup,
-      100213 => :other_makeup_products,
-      100221 => :perfumes,
-      100222 => :hydroalcoholic_perfumes,
-      100225 => :non_hydroalcoholic_perfumes,
-      100230 => :sun_and_selftanning_products,
-      100231 => :before_and_after_sun_products,
-      100239 => :sun_protection_products,
-      100246 => :selftanning_products,
-      100250 => :other_sun_and_selftanning_products,
-      100252 => :other_skin_products,
-      100253 => :other_skin_products_child,
-      100255 => :hair_and_scalp_products,
-      100256 => :hair_and_scalp_care_and_cleansing_products,
-      100257 => :shampoo,
-      100262 => :hair_conditioner,
-      100268 => :scalp_and_hair_roots_care_products,
-      100272 => :antidandruff_products,
-      100278 => :antihairloss_products,
-      100282 => :other_hair_and_scalp_care_and_cleansing_products,
-      100284 => :hair_colouring_products,
-      100285 => :oxidative_hair_colour_products,
-      100293 => :nonoxidative_hair_colour_products,
-      100298 => :hair_bleaching_and_dye_remover_products,
-      100310 => :other_hair_colouring_products,
-      100312 => :hair_styling_products,
-      100313 => :products_for_temporary_hair_styling,
-      100324 => :permanent_wave_products,
-      100332 => :hair_relaxer_straightener_products,
-      100338 => :other_hair_styling_products,
-      100340 => :other_hair_and_scalp_products,
-      100341 => :hair_sun_protection_products,
-      100347 => :other_hair_and_scalp_products_child,
-      100349 => :nail_and_cuticle_products,
-      100350 => :nail_varnish_and_remover_products,
-      100351 => :nail_varnish_nail_makeup,
-      100355 => :nail_varnish_remover,
-      100357 => :nail_varnish_thinner,
-      100359 => :nail_bleach,
-      100361 => :other_nail_varnish_and_remover_products,
-      100363 => :nail_care_nail_hardener_products,
-      100364 => :nail_care_products,
-      100367 => :nail_hardener,
-      100369 => :other_nail_care_nail_hardener_products,
-      100371 => :nail_glue_remover_products,
-      100372 => :nail_glue_remover,
-      100374 => :other_nail_and_cuticle_products,
-      100375 => :cuticle_remover_softener,
-      100377 => :nail_sculpting_products,
-      100379 => :other_nail_and_cuticle_products_child,
-      100381 => :oral_hygiene_products,
-      100382 => :tooth_care_products,
-      100383 => :toothpaste,
-      100386 => :tooth_cleansing_powder_salt,
-      100388 => :other_tooth_care_products,
-      100390 => :mouth_wash_breath_spray,
-      100391 => :mouth_wash,
-      100395 => :breath_spray,
-      100398 => :other_mouth_wash_breath_spray_products,
-      100400 => :tooth_whiteners,
-      100401 => :tooth_whiteners_child,
-      100403 => :other_oral_hygiene_products,
-      100404 => :other_oral_hygiene_products_child
-  }.freeze
 
   FRAME_FORMULATION_ID = {
       100006 => :skin_care_cream_lotion_gel,
@@ -361,81 +228,5 @@ module CpnpNotificationInfo
       100406 => :dry_shampoo,
       100407 => :syndet_bar_soap,
       100408 => :combo_bar_soap
-  }.freeze
-
-  TRIGGER_RULES_QUESTION_ID = {
-      100026 => :please_specify_the_percentage_weight_of_ethanol,
-      100027 => :please_specify_the_percentage_weight_of_isopropanol,
-      100001 => :please_specify_the_inci_name_and_concentration_of_the_antidandruff_agents_if_antidandruff_agents_are_not_present_in_the_cosmetic_product_then_not_applicable_must_be_checked,
-      100008 => :please_specify_the_inci_name_and_concentration_of_the_antihair_loss_agents_if_antihair_loss_agents_are_not_present_in_the_cosmetic_product_then_not_applicable_must_be_checked,
-      100009 => :please_specify_the_inci_name_and_concentration_of_the_antipigmenting_and_depigmenting_agents_if_antipigmenting_and_depigmenting_agents_are_not_present_in_the_cosmetic_product_then_not_applicable_must_be_checked,
-      100010 => :please_specify_the_inci_name_and_concentration_of_chemical_exfoliating_agents_if_chemical_exfoliating_agents_are_not_present_in_the_cosmetic_product_then_not_applicable_must_be_checked,
-      100011 => :please_specify_the_exact_content_of_vitamin_a_or_its_derivatives_for_the_whole_product_if_the_level_of_vitamin_a_or_any_of_its_derivatives_does_not_exceed_020_calculated_as_retinol_or_if_the_amount_does_not_exceed_009_grams_calculated_as_retinol_or_if_vitamin_a_or_any_of_its_derivatives_are_not_present_in_the_product_then_not_applicable_must_be_checked,
-      100012 => :please_specify_the_inci_name_and_the_concentration_of_xanthine_derivatives_eg_caffeine_theophylline_theobromine_plant_extracts_containing_xanthine_derivatives_eg_paulinia_cupana_guarana_extractspowders_if_xanthine_derivatives_are_not_present_or_present_below_05_in_the_cosmetic_product_then_not_applicable_must_be_checked,
-      100013 => :please_specify_the_inci_name_and_concentration_of_the_cationic_surfactants_with_two_or_more_chain_lengths_below_c12_if_the_surfactant_is_used_for_non_preservative_purpose_if_cationic_surfactants_with_two_or_more_chain_lengths_below_c12_are_not_present_in_the_product_then_not_applicable_must_be_checked,
-      100014 => :please_specify_the_inci_name_and_concentration_of_each_propellant_if_propellants_are_not_present_in_the_product_then_not_applicable_must_be_checked,
-      100015 => :please_specify_the_concentration_of_hydrogen_peroxide_if_hydrogen_peroxide_is_not_present_in_the_product_then_not_applicable_must_be_checked_,
-      100016 => :please_specify_the_inci_name_and_the_concentration_of_the_compounds_that_release_hydrogen_peroxide_if_compounds_releasing_hydrogen_peroxide_are_not_present_in_the_product_then_not_applicable_must_be_checked,
-      100017 => :please_specify_the_inci_name_and_concentration_of_each_reducing_agent_if_reducing_agents_are_not_present_in_the_product_then_not_applicable_must_be_checked,
-      100018 => :please_specify_the_inci_name_and_concentration_of_each_persulfate_if_persulfates_are_not_present_in_the_product_then_not_applicable_must_be_checked,
-      100019 => :please_specify_the_inci_name_and_concentration_of_each_straightening_agent_if_straightening_agents_are_not_present_in_the_product_then_not_applicable_must_be_checked,
-      100020 => :please_indicate_the_total_concentration_of_inorganic_sodium_salts_if_inorganic_sodium_salts_are_not_present_in_the_product_then_not_applicable_must_be_checked,
-      100021 => :please_indicate_the_concentration_of_fluoride_compounds_calculated_as_fluorine_if_fluoride_compounds_are_not_present_in_the_product_then_not_applicable_must_be_checked,
-      100003 => :is_the_ph_of_the_component_lower_than_3_or_higher_than_10,
-      100024 => :please_indicate_the_total_level_of_essential_oils_camphor_menthol_or_eucalyptol_if_essential_oils_camphor_menthol_or_eucalyptol_are_not_present_in_the_product_or_if_the_level_of_essential_oils_camphor_menthol_or_eucalyptol_does_not_exceed_05_then_not_applicable_must_be_checked,
-      100025 => :please_indicate_the_name_and_the_quantity_of_each_essential_oil_camphor_menthol_or_eucalyptol_if_no_individual_essential_oil_camphor_menthol_or_eucalyptol_are_present_with_a_level_higher_than_05_015_in_case_of_camphor_then_not_applicable_must_be_checked,
-      100002 => :please_indicate_the_total_concentration_of_inorganic_sodium_salts,
-      100023 => :please_indicate_the_ph_of_the_hair_dye_component,
-      100022 => :please_indicate_the_ph_of_the_mixed_hair_dye_product,
-      100004 => :please_indicate_the_ph,
-      100007 => :please_indicate_the_ph_of_the_mixed_product_,
-      100006 => :do_the_components_of_the_product_need_to_be_mixed,
-      100005 => :please_indicate_the_inci_name_and_concentration_of_each_alkaline_agent_including_ammonium_hydroxide_liberators
-  }.freeze
-
-  TRIGGER_RULES_QUESTION_ELEMENT_ID = {
-      100042 => :ethanol,
-      100043 => :propanol,
-      100001 => :inciname,
-      100002 => :incivalue,
-      100011 => :inciname,
-      100012 => :incivalue,
-      100013 => :inciname,
-      100014 => :incivalue,
-      100015 => :inciname,
-      100016 => :incivalue,
-      100017 => :value,
-      100018 => :inciname,
-      100019 => :incivalue,
-      100020 => :inciname,
-      100021 => :incivalue,
-      100022 => :inciname,
-      100023 => :incivalue,
-      100024 => :value,
-      100025 => :inciname,
-      100026 => :incivalue,
-      100027 => :inciname,
-      100028 => :incivalue,
-      100029 => :inciname,
-      100030 => :incivalue,
-      100031 => :inciname,
-      100032 => :incivalue,
-      100033 => :value,
-      100034 => :value,
-      100004 => :ph,
-      100039 => :value,
-      100040 => :inciname,
-      100041 => :incivalue,
-      100003 => :concentration,
-      100037 => :minrangevalue,
-      100038 => :maxrangevalue,
-      100035 => :minrangevalue,
-      100036 => :maxrangevalue,
-      100005 => :ph,
-      100009 => :minrangevalue,
-      100010 => :maxrangevalue,
-      100008 => :ph,
-      100006 => :inciname,
-      100007 => :incivalue
   }.freeze
 end
