+++ cosmetics-web/app/models/concerns/cpnp_notification_properties.rb
@@ -1,0 +1,7 @@
+module CpnpNotificationProperties
+  include CpnpNotificationTypes
+  include CpnpNotificationCategories
+  include CpnpNotificationFrameFormulations
+  include CpnpNotificationTriggerRules
+  include CpnpNotificationUnits
+end
