+++ cosmetics-web/app/models/concerns/notification_properties.rb
@@ -1,0 +1,9 @@
+module NotificationProperties
+  extend ActiveSupport::Concern
+
+  include NotificationTypes
+  include NotificationCategories
+  include NotificationFrameFormulations
+  include NotificationTriggerRules
+  include NotificationUnits
+end
