+++ cosmetics-web/app/models/concerns/notification_types.rb
@@ -1,0 +1,11 @@
+module NotificationTypes
+  extend ActiveSupport::Concern
+
+  included do
+    enum notification_type: {
+        predefined: "predefined",
+        exact: "exact",
+        range: "range"
+    }
+  end
+end
