--- cosmetics-web/app/models/concerns/uk_notification_info.rb
@@ -1,500 +1,0 @@
-# rubocop:disable Metrics/BlockLength
-module UkNotificationInfo
-  extend ActiveSupport::Concern
-
-  included do
-    enum notification_type: {
-        predefined: "predefined",
-        exact: "exact",
-        range: "range"
-    }
-
-    enum category: {
-        skin_products: "skin_products",
-        skin_care_products: "skin_care_products",
-        face_care_products_other_than_face_mask: "face_care_products_other_than_face_mask",
-        face_mask: "face_mask",
-        eye_contour_products: "eye_contour_products",
-        lip_care_products: "lip_care_products",
-        hand_care_products: "hand_care_products",
-        foot_care_products: "foot_care_products",
-        body_care_products: "body_care_products",
-        external_intimate_care_products: "external_intimate_care_products",
-        chemical_exfoliation_products: "chemical_exfoliation_products",
-        mechanical_exfoliation_products: "mechanical_exfoliation_products",
-        skin_lightening_products: "skin_lightening_products",
-        other_skin_care_products: "other_skin_care_products",
-        skin_cleansing_products: "skin_cleansing_products",
-        soap_products: "soap_products",
-        bath_shower_products: "bath_shower_products",
-        makeup_remover_products: "makeup_remover_products",
-        external_intimate_hygiene_products: "external_intimate_hygiene_products",
-        other_skin_cleansing_products: "other_skin_cleansing_products",
-        body_hair_removal_products: "body_hair_removal_products",
-        chemical_depilatories: "chemical_depilatories",
-        physical_epilation_products: "physical_epilation_products",
-        other_body_hair_removal_products: "other_body_hair_removal_products",
-        bleach_for_body_hair_products: "bleach_for_body_hair_products",
-        bleach_for_body_hair: "bleach_for_body_hair",
-        correction_of_body_odour_andor_perspiration: "correction_of_body_odour_andor_perspiration",
-        products_with_antiperspirant_activity: "products_with_antiperspirant_activity",
-        products_without_antiperspirant_activity: "products_without_antiperspirant_activity",
-        shaving_and_pre_after_shaving_products: "shaving_and_pre_after_shaving_products",
-        shaving_products: "shaving_products",
-        pre_aftershaving_products: "pre_aftershaving_products",
-        other_shaving_and_pre_after_shaving_products: "other_shaving_and_pre_after_shaving_products",
-        makeup_products: "makeup_products",
-        foundation: "foundation",
-        concealer: "concealer",
-        other_face_makeup_products: "other_face_makeup_products",
-        mascara: "mascara",
-        eye_shadow: "eye_shadow",
-        eye_pencil: "eye_pencil",
-        eye_liner: "eye_liner",
-        other_eye_makeup_products: "other_eye_makeup_products",
-        lip_stick: "lip_stick",
-        lipstick_sealer: "lipstick_sealer",
-        other_lip_makeup_products: "other_lip_makeup_products",
-        body_or_face_paint_including_carneval_makeup: "body_or_face_paint_including_carneval_makeup",
-        other_makeup_products: "other_makeup_products",
-        perfumes: "perfumes",
-        hydroalcoholic_perfumes: "hydroalcoholic_perfumes",
-        non_hydroalcoholic_perfumes: "non_hydroalcoholic_perfumes",
-        sun_and_selftanning_products: "sun_and_selftanning_products",
-        before_and_after_sun_products: "before_and_after_sun_products",
-        sun_protection_products: "sun_protection_products",
-        selftanning_products: "selftanning_products",
-        other_sun_and_selftanning_products: "other_sun_and_selftanning_products",
-        other_skin_products: "other_skin_products",
-        other_skin_products_child: "other_skin_products_child",
-        hair_and_scalp_products: "hair_and_scalp_products",
-        hair_and_scalp_care_and_cleansing_products: "hair_and_scalp_care_and_cleansing_products",
-        shampoo: "shampoo",
-        hair_conditioner: "hair_conditioner",
-        scalp_and_hair_roots_care_products: "scalp_and_hair_roots_care_products",
-        antidandruff_products: "antidandruff_products",
-        antihairloss_products: "antihairloss_products",
-        other_hair_and_scalp_care_and_cleansing_products: "other_hair_and_scalp_care_and_cleansing_products",
-        hair_colouring_products: "hair_colouring_products",
-        oxidative_hair_colour_products: "oxidative_hair_colour_products",
-        nonoxidative_hair_colour_products: "nonoxidative_hair_colour_products",
-        hair_bleaching_and_dye_remover_products: "hair_bleaching_and_dye_remover_products",
-        other_hair_colouring_products: "other_hair_colouring_products",
-        hair_styling_products: "hair_styling_products",
-        products_for_temporary_hair_styling: "products_for_temporary_hair_styling",
-        permanent_wave_products: "permanent_wave_products",
-        hair_relaxer_straightener_products: "hair_relaxer_straightener_products",
-        other_hair_styling_products: "other_hair_styling_products",
-        other_hair_and_scalp_products: "other_hair_and_scalp_products",
-        hair_sun_protection_products: "hair_sun_protection_products",
-        other_hair_and_scalp_products_child: "other_hair_and_scalp_products_child",
-        nail_and_cuticle_products: "nail_and_cuticle_products",
-        nail_varnish_and_remover_products: "nail_varnish_and_remover_products",
-        nail_varnish_nail_makeup: "nail_varnish_nail_makeup",
-        nail_varnish_remover: "nail_varnish_remover",
-        nail_varnish_thinner: "nail_varnish_thinner",
-        nail_bleach: "nail_bleach",
-        other_nail_varnish_and_remover_products: "other_nail_varnish_and_remover_products",
-        nail_care_nail_hardener_products: "nail_care_nail_hardener_products",
-        nail_care_products: "nail_care_products",
-        nail_hardener: "nail_hardener",
-        other_nail_care_nail_hardener_products: "other_nail_care_nail_hardener_products",
-        nail_glue_remover_products: "nail_glue_remover_products",
-        nail_glue_remover: "nail_glue_remover",
-        other_nail_and_cuticle_products: "other_nail_and_cuticle_products",
-        cuticle_remover_softener: "cuticle_remover_softener",
-        nail_sculpting_products: "nail_sculpting_products",
-        other_nail_and_cuticle_products_child: "other_nail_and_cuticle_products_child",
-        oral_hygiene_products: "oral_hygiene_products",
-        tooth_care_products: "tooth_care_products",
-        toothpaste: "toothpaste",
-        tooth_cleansing_powder_salt: "tooth_cleansing_powder_salt",
-        other_tooth_care_products: "other_tooth_care_products",
-        mouth_wash_breath_spray: "mouth_wash_breath_spray",
-        mouth_wash: "mouth_wash",
-        breath_spray: "breath_spray",
-        other_mouth_wash_breath_spray_products: "other_mouth_wash_breath_spray_products",
-        tooth_whiteners: "tooth_whiteners",
-        tooth_whiteners_child: "tooth_whiteners_child",
-        other_oral_hygiene_products: "other_oral_hygiene_products",
-        other_oral_hygiene_products_child: "other_oral_hygiene_products_child"
-    }, _prefix: :category
-
-    enum frame_formulation: {
-        skin_care_cream_lotion_gel: "skin_care_cream_lotion_gel",
-        skin_care_cream_lotion_gel_with_high_level_of_silicones: "skin_care_cream_lotion_gel_with_high_level_of_silicones",
-        skin_care_cream_lotion_gel_with_high_level_of_humectants: "skin_care_cream_lotion_gel_with_high_level_of_humectants",
-        skin_care_cream_lotion_gel_with_high_level_of_bulking_agents: "skin_care_cream_lotion_gel_with_high_level_of_bulking_agents",
-        skin_care_cream_lotion_gel_with_high_level_of_fatty_compounds: "skin_care_cream_lotion_gel_with_high_level_of_fatty_compounds",
-        skin_care_cream_lotion_gel_with_high_level_of_uv_filters: "skin_care_cream_lotion_gel_with_high_level_of_uv_filters",
-        skin_gel_hydroalcoholic_based: "skin_gel_hydroalcoholic_based",
-        skin_oil: "skin_oil",
-        tonic_cosmetic_water: "tonic_cosmetic_water",
-        skin_care_silicone_based: "skin_care_silicone_based",
-        petroleum_jelly: "petroleum_jelly",
-        other_face_care_products_other_than_face_mask: "other_face_care_products_other_than_face_mask",
-        other_face_mask_products: "other_face_mask_products",
-        other_eye_contour_products: "other_eye_contour_products",
-        lip_care: "lip_care",
-        other_lip_care_products: "other_lip_care_products",
-        other_hand_care_products: "other_hand_care_products",
-        other_foot_care_products: "other_foot_care_products",
-        skin_care_cream_lotion_gel_with_high_level_of_perfume: "skin_care_cream_lotion_gel_with_high_level_of_perfume",
-        zinc_oxide_based_cream: "zinc_oxide_based_cream",
-        talcum_powder_powder_or_pressed: "talcum_powder_powder_or_pressed",
-        other_body_care_products: "other_body_care_products",
-        external_intimate_care_products: "external_intimate_care_products",
-        chemical_exfoliation_products: "chemical_exfoliation_products",
-        other_chemical_exfoliation_products: "other_chemical_exfoliation_products",
-        skin_cleansing_cream_lotion_scrub_gel: "skin_cleansing_cream_lotion_scrub_gel",
-        body_scrub_gel_cream: "body_scrub_gel_cream",
-        other_mechnical_exfoliation_products: "other_mechnical_exfoliation_products",
-        other_skin_lightening_products: "other_skin_lightening_products",
-        foaming_skin_cleanser: "foaming_skin_cleanser",
-        foaming_skin_cleanser_with_high_level_of_oils_andor_humectants: "foaming_skin_cleanser_with_high_level_of_oils_andor_humectants",
-        gentle_cleansing_gel: "gentle_cleansing_gel",
-        other_skin_care_products: "other_skin_care_products",
-        soap_toilet: "soap_toilet",
-        liquid_soap: "liquid_soap",
-        other_soap_products: "other_soap_products",
-        bath_and_shower_gel_cream_milk_paste: "bath_and_shower_gel_cream_milk_paste",
-        slightly_foaming_shower_cream_milk: "slightly_foaming_shower_cream_milk",
-        bath_salts_cubes: "bath_salts_cubes",
-        bath_and_shower_oil_including_bath_beads: "bath_and_shower_oil_including_bath_beads",
-        footbath: "footbath",
-        other_bath_shower_products: "other_bath_shower_products",
-        makeup_remover_non_foaming_cleanser_including_twophase_products: "makeup_remover_non_foaming_cleanser_including_twophase_products",
-        makeup_remover_non_foaming_oil_cleanser: "makeup_remover_non_foaming_oil_cleanser",
-        other_makeup_remover_products: "other_makeup_remover_products",
-        other_external_intimate_hygiene_products: "other_external_intimate_hygiene_products",
-        other_skin_cleansing_products: "other_skin_cleansing_products",
-        chemical_depilatory_cream_gel: "chemical_depilatory_cream_gel",
-        other_chemical_depilatory_products: "other_chemical_depilatory_products",
-        physical_depilation_product_wax_resin_based: "physical_depilation_product_wax_resin_based",
-        physical_depilation_product_wax_sugar_based: "physical_depilation_product_wax_sugar_based",
-        other_physical_epilation_products: "other_physical_epilation_products",
-        other_body_hair_removal_products: "other_body_hair_removal_products",
-        bleach_for_body_hair: "bleach_for_body_hair",
-        other_bleach_for_body_hair_products: "other_bleach_for_body_hair_products",
-        antiperspirant_aerosol_spray_with_antiperspirant_salt: "antiperspirant_aerosol_spray_with_antiperspirant_salt",
-        antiperspirant_rollon_with_antiperspirant_salt: "antiperspirant_rollon_with_antiperspirant_salt",
-        antiperspirant_stick_with_antiperspirant_salt: "antiperspirant_stick_with_antiperspirant_salt",
-        antiperspirant_liquid_squeeze_pack_or_pump_spray: "antiperspirant_liquid_squeeze_pack_or_pump_spray",
-        antiperspirant_cream: "antiperspirant_cream",
-        other_products_with_antiperspirant_activity: "other_products_with_antiperspirant_activity",
-        deodorant_aerosol_spray_without_antiperspirant_salt: "deodorant_aerosol_spray_without_antiperspirant_salt",
-        deodorant_rollon_without_antiperspirant_salt: "deodorant_rollon_without_antiperspirant_salt",
-        deodorant_stick_without_antiperspirant_salt: "deodorant_stick_without_antiperspirant_salt",
-        deodorant_liquid_squeeze_pack_or_pump_spray: "deodorant_liquid_squeeze_pack_or_pump_spray",
-        other_products_without_antiperspirant_activity: "other_products_without_antiperspirant_activity",
-        shaving_cream: "shaving_cream",
-        brushless_shaving_cream: "brushless_shaving_cream",
-        shave_gel: "shave_gel",
-        shaving_gel_post_foaming: "shaving_gel_post_foaming",
-        shaving_foam_aerosol: "shaving_foam_aerosol",
-        shaving_soapstick: "shaving_soapstick",
-        other_shaving_and_pre_after_shaving_products: "other_shaving_and_pre_after_shaving_products",
-        aftershave_balm: "aftershave_balm",
-        aftershave_lotion: "aftershave_lotion",
-        other_pre_aftershaving_products: "other_pre_aftershaving_products",
-        foundation_and_blusher_liquid_cream_foam: "foundation_and_blusher_liquid_cream_foam",
-        foundation_anhydrous: "foundation_anhydrous",
-        foundation_concealer_solid: "foundation_concealer_solid",
-        other_foundation_products: "other_foundation_products",
-        concealer_cream: "concealer_cream",
-        concealer_stick: "concealer_stick",
-        lipstick_lip_cream_including_concealer_stickcream: "lipstick_lip_cream_including_concealer_stickcream",
-        other_concealer_products: "other_concealer_products",
-        other_face_makeup_products: "other_face_makeup_products",
-        mascara_regular: "mascara_regular",
-        mascara_waterproof: "mascara_waterproof",
-        mascara_cake: "mascara_cake",
-        other_mascara_products: "other_mascara_products",
-        eye_shadow_blusher_and_liner_powder: "eye_shadow_blusher_and_liner_powder",
-        eye_shadow_and_blusher_cream: "eye_shadow_and_blusher_cream",
-        eye_shadow_and_blusher_anhydrous_cream_stick_powder_: "eye_shadow_and_blusher_anhydrous_cream_stick_powder_",
-        other_eye_shadow_products: "other_eye_shadow_products",
-        eye_pencil: "eye_pencil",
-        other_eye_pencil_products: "other_eye_pencil_products",
-        eye_liner_regular_liquid_cream: "eye_liner_regular_liquid_cream",
-        eye_liner_waterproof_liquid_cream_paste: "eye_liner_waterproof_liquid_cream_paste",
-        eye_liner_cake: "eye_liner_cake",
-        other_eye_liner_products: "other_eye_liner_products",
-        lip_and_eye_fix_regular: "lip_and_eye_fix_regular",
-        eye_fix_anhydrous: "eye_fix_anhydrous",
-        other_eye_makeup_products: "other_eye_makeup_products",
-        other_lip_stick_products: "other_lip_stick_products",
-        lipstick_sealer_lip_coat: "lipstick_sealer_lip_coat",
-        other_lipstick_sealer_products: "other_lipstick_sealer_products",
-        lip_colour_and_glosserbalm: "lip_colour_and_glosserbalm",
-        lip_pencil: "lip_pencil",
-        lip_ink_pencil: "lip_ink_pencil",
-        lip_ink_liquid: "lip_ink_liquid",
-        other_lip_makeup_products: "other_lip_makeup_products",
-        body_glitter: "body_glitter",
-        other_body_or_face_paint_products: "other_body_or_face_paint_products",
-        face_powder: "face_powder",
-        other_makeup_products: "other_makeup_products",
-        perfume_eau_de_cologne_eau_de_toilette_eau_de_parfum_extrait_de_parfum_parfum: "perfume_eau_de_cologne_eau_de_toilette_eau_de_parfum_extrait_de_parfum_parfum",
-        other_hydroalcoholic_perfumes: "other_hydroalcoholic_perfumes",
-        perfumed_powder_loose_or_pressed: "perfumed_powder_loose_or_pressed",
-        solid_perfume_paste_waxe_: "solid_perfume_paste_waxe_",
-        perfumed_skin_oil: "perfumed_skin_oil",
-        other_non_hydroalcoholic_perfumes: "other_non_hydroalcoholic_perfumes",
-        other_before_and_after_sun_products: "other_before_and_after_sun_products",
-        sunscreen_cream_lotion: "sunscreen_cream_lotion",
-        sunscreen_oil: "sunscreen_oil",
-        sunscreen_stick: "sunscreen_stick",
-        sunscreen_hydrogel: "sunscreen_hydrogel",
-        sunscreen_gel_oil_based: "sunscreen_gel_oil_based",
-        other_sun_protection_products: "other_sun_protection_products",
-        self_tan_lotion_cream: "self_tan_lotion_cream",
-        self_tan_lotion_gel: "self_tan_lotion_gel",
-        other_selftanning_products: "other_selftanning_products",
-        other_sun_and_selftanning_products: "other_sun_and_selftanning_products",
-        other_skin_products: "other_skin_products",
-        shampoo_liquid_cream_paste: "shampoo_liquid_cream_paste",
-        soap_shampoo: "soap_shampoo",
-        shampoo_plus_conditioner: "shampoo_plus_conditioner",
-        other_shampoo_products: "other_shampoo_products",
-        hair_conditioner: "hair_conditioner",
-        hair_conditioner_silicone_based: "hair_conditioner_silicone_based",
-        hair_scalp_care_lotion: "hair_scalp_care_lotion",
-        hair_scalp_biphase_lotion: "hair_scalp_biphase_lotion",
-        other_hair_conditioner_products: "other_hair_conditioner_products",
-        other_scalp_and_hair_roots_care_products: "other_scalp_and_hair_roots_care_products",
-        shampoo_liquid_cream_paste_child: "shampoo_liquid_cream_paste_child",
-        other_antidandruff_products: "other_antidandruff_products",
-        other_antihairloss_products: "other_antihairloss_products",
-        other_hair_and_scalp_care_and_cleansing_products: "other_hair_and_scalp_care_and_cleansing_products",
-        hair_colorant_permanent_oxidative_type_type_1_two_components_colorant_part: "hair_colorant_permanent_oxidative_type_type_1_two_components_colorant_part",
-        hair_colorant_permanent_oxidative_type_type_2_two_components_colorant_part_with_high_level_of_oils: "hair_colorant_permanent_oxidative_type_type_2_two_components_colorant_part_with_high_level_of_oils",
-        hair_colorant_permanent_oxidative_type_type_3_three_components_colorant_part_a: "hair_colorant_permanent_oxidative_type_type_3_three_components_colorant_part_a",
-        hair_colorant_permanent_oxidative_type_type_3_three_components_colorant_part_b: "hair_colorant_permanent_oxidative_type_type_3_three_components_colorant_part_b",
-        hair_colorant_permanent_oxidative_type_type_1_two_or_three_components_oxidative_part: "hair_colorant_permanent_oxidative_type_type_1_two_or_three_components_oxidative_part",
-        hair_colorant_permanent_oxidative_type_type_2_two_or_three_components_oxidative_part_with_high_level_of_oils: "hair_colorant_permanent_oxidative_type_type_2_two_or_three_components_oxidative_part_with_high_level_of_oils",
-        other_oxidative_hair_colour_products: "other_oxidative_hair_colour_products",
-        hair_colorant_temporary_shampoo: "hair_colorant_temporary_shampoo",
-        hair_colorant_temporary_foam_lotion: "hair_colorant_temporary_foam_lotion",
-        hair_colorant_temporary_or_semipermanent_liquid_cream_foam: "hair_colorant_temporary_or_semipermanent_liquid_cream_foam",
-        other_nonoxidative_hair_colour_products: "other_nonoxidative_hair_colour_products",
-        dye_remover_reducing_type_part_a: "dye_remover_reducing_type_part_a",
-        dye_remover_reducing_type_part_b: "dye_remover_reducing_type_part_b",
-        dye_remover_oxidative_type_part_a: "dye_remover_oxidative_type_part_a",
-        dye_remover_oxidative_type_part_b: "dye_remover_oxidative_type_part_b",
-        skin_dyeremoving_cleanser_reducing_type: "skin_dyeremoving_cleanser_reducing_type",
-        hair_bleach_powder: "hair_bleach_powder",
-        hair_bleach_paste: "hair_bleach_paste",
-        hair_bleach_oxidative_part_type_1: "hair_bleach_oxidative_part_type_1",
-        hair_bleach_oxidative_part_type_2: "hair_bleach_oxidative_part_type_2",
-        hair_bleach_alkaline_part: "hair_bleach_alkaline_part",
-        other_hair_bleaching_and_dye_remover_products: "other_hair_bleaching_and_dye_remover_products",
-        other_hair_colouring_products: "other_hair_colouring_products",
-        hair_styling_cream_paste_gum_pomade_paste_waxe_: "hair_styling_cream_paste_gum_pomade_paste_waxe_",
-        hair_shine_wax_or_brilliantine: "hair_shine_wax_or_brilliantine",
-        hair_styling_lotion: "hair_styling_lotion",
-        setting_lotion_including_coloured_setting_lotions: "setting_lotion_including_coloured_setting_lotions",
-        hair_shine_styling_gel: "hair_shine_styling_gel",
-        hair_foam: "hair_foam",
-        hair_spray_aerosol: "hair_spray_aerosol",
-        hair_spray_pump_spray: "hair_spray_pump_spray",
-        hair_shine_spray_aerosol: "hair_shine_spray_aerosol",
-        other_products_for_temporary_hair_styling: "other_products_for_temporary_hair_styling",
-        permanent_wave_lotion_single_component_based_on_reducing_agents: "permanent_wave_lotion_single_component_based_on_reducing_agents",
-        permanent_wave_lotion_twocomponent_based_on_hydrogen_peroxide_and_reducing_agents_part_a: "permanent_wave_lotion_twocomponent_based_on_hydrogen_peroxide_and_reducing_agents_part_a",
-        permanent_wave_lotion_twocomponent_based_on_hydrogen_peroxide_and_reducing_agents_part_b: "permanent_wave_lotion_twocomponent_based_on_hydrogen_peroxide_and_reducing_agents_part_b",
-        permanent_wave_lotion_twocomponent_based_on_reducing_agents_part_a: "permanent_wave_lotion_twocomponent_based_on_reducing_agents_part_a",
-        permanent_wave_lotion_twocomponent_based_on_reducing_agents_part_b: "permanent_wave_lotion_twocomponent_based_on_reducing_agents_part_b",
-        permanent_wave_neutraliser_hair_straightener_relaxer_neutraliser_liquid_cream: "permanent_wave_neutraliser_hair_straightener_relaxer_neutraliser_liquid_cream",
-        other_permanent_wave_products: "other_permanent_wave_products",
-        hair_straightener_relaxer_type_1: "hair_straightener_relaxer_type_1",
-        hair_straightener_relaxer_type_2_part_a: "hair_straightener_relaxer_type_2_part_a",
-        hair_straightener_relaxer_type_2_part_b: "hair_straightener_relaxer_type_2_part_b",
-        hair_straightener_relaxer_type_3: "hair_straightener_relaxer_type_3",
-        other_hair_relaxer_straightener_products: "other_hair_relaxer_straightener_products",
-        other_hair_styling_products: "other_hair_styling_products",
-        other_hair_sun_protection_products: "other_hair_sun_protection_products",
-        other_hair_and_scalp_products: "other_hair_and_scalp_products",
-        nail_varnish: "nail_varnish",
-        nail_white_pencil: "nail_white_pencil",
-        other_nail_varnish_nail_makeup: "other_nail_varnish_nail_makeup",
-        nail_varnish_remover: "nail_varnish_remover",
-        nail_varnish_thinner: "nail_varnish_thinner",
-        nail_bleach: "nail_bleach",
-        other_nail_varnish_and_remover_products: "other_nail_varnish_and_remover_products",
-        nail_conditioner_oil_based: "nail_conditioner_oil_based",
-        other_nail_care_products: "other_nail_care_products",
-        nail_hardener: "nail_hardener",
-        other_nail_care_nail_hardener_products: "other_nail_care_nail_hardener_products",
-        nail_glue_remover: "nail_glue_remover",
-        cuticle_remover_softener: "cuticle_remover_softener",
-        nail_sculpting_products: "nail_sculpting_products",
-        other_nail_and_cuticle_products: "other_nail_and_cuticle_products",
-        toothpaste: "toothpaste",
-        other_toothpaste_products: "other_toothpaste_products",
-        tooth_cleansing_powder_salt: "tooth_cleansing_powder_salt",
-        other_tooth_care_products: "other_tooth_care_products",
-        mouthwash: "mouthwash",
-        mouthwash_concentrate: "mouthwash_concentrate",
-        other_mouth_wash_products: "other_mouth_wash_products",
-        breath_spray_pump_spray: "breath_spray_pump_spray",
-        others_breath_spray_products: "others_breath_spray_products",
-        other_mouth_wash_breath_spray_products: "other_mouth_wash_breath_spray_products",
-        tooth_whiteners: "tooth_whiteners",
-        other_oral_hygiene_products: "other_oral_hygiene_products",
-        dry_shampoo: "dry_shampoo",
-        syndet_bar_soap: "syndet_bar_soap",
-        combo_bar_soap: "combo_bar_soap"
-    }, _prefix: :frame_formulation
-
-    PARENT_OF_CATEGORY = {
-        skin_care_products: :skin_products,
-        face_care_products_other_than_face_mask: :skin_care_products,
-        face_mask: :skin_care_products,
-        eye_contour_products: :skin_care_products,
-        lip_care_products: :skin_care_products,
-        hand_care_products: :skin_care_products,
-        foot_care_products: :skin_care_products,
-        body_care_products: :skin_care_products,
-        external_intimate_care_products: :skin_care_products,
-        chemical_exfoliation_products: :skin_care_products,
-        mechanical_exfoliation_products: :skin_care_products,
-        skin_lightening_products: :skin_care_products,
-        other_skin_care_products: :skin_care_products,
-        skin_cleansing_products: :skin_products,
-        soap_products: :skin_cleansing_products,
-        bath_shower_products: :skin_cleansing_products,
-        makeup_remover_products: :skin_cleansing_products,
-        external_intimate_hygiene_products: :skin_cleansing_products,
-        other_skin_cleansing_products: :skin_cleansing_products,
-        body_hair_removal_products: :skin_products,
-        chemical_depilatories: :body_hair_removal_products,
-        physical_epilation_products: :body_hair_removal_products,
-        other_body_hair_removal_products: :body_hair_removal_products,
-        bleach_for_body_hair_products: :skin_products,
-        bleach_for_body_hair: :bleach_for_body_hair_products,
-        correction_of_body_odour_andor_perspiration: :skin_products,
-        products_with_antiperspirant_activity: :correction_of_body_odour_andor_perspiration,
-        products_without_antiperspirant_activity: :correction_of_body_odour_andor_perspiration,
-        shaving_and_pre_after_shaving_products: :skin_products,
-        shaving_products: :shaving_and_pre_after_shaving_products,
-        pre_aftershaving_products: :shaving_and_pre_after_shaving_products,
-        other_shaving_and_pre_after_shaving_products: :shaving_and_pre_after_shaving_products,
-        makeup_products: :skin_products,
-        foundation: :makeup_products,
-        concealer: :makeup_products,
-        other_face_makeup_products: :makeup_products,
-        mascara: :makeup_products,
-        eye_shadow: :makeup_products,
-        eye_pencil: :makeup_products,
-        eye_liner: :makeup_products,
-        other_eye_makeup_products: :makeup_products,
-        lip_stick: :makeup_products,
-        lipstick_sealer: :makeup_products,
-        other_lip_makeup_products: :makeup_products,
-        body_or_face_paint_including_carneval_makeup: :makeup_products,
-        other_makeup_products: :makeup_products,
-        perfumes: :skin_products,
-        hydroalcoholic_perfumes: :perfumes,
-        non_hydroalcoholic_perfumes: :perfumes,
-        sun_and_selftanning_products: :skin_products,
-        before_and_after_sun_products: :sun_and_selftanning_products,
-        sun_protection_products: :sun_and_selftanning_products,
-        selftanning_products: :sun_and_selftanning_products,
-        other_sun_and_selftanning_products: :sun_and_selftanning_products,
-        other_skin_products: :skin_products,
-        other_skin_products_child: :other_skin_products,
-        hair_and_scalp_care_and_cleansing_products: :hair_and_scalp_products,
-        shampoo: :hair_and_scalp_care_and_cleansing_products,
-        hair_conditioner: :hair_and_scalp_care_and_cleansing_products,
-        scalp_and_hair_roots_care_products: :hair_and_scalp_care_and_cleansing_products,
-        antidandruff_products: :hair_and_scalp_care_and_cleansing_products,
-        antihairloss_products: :hair_and_scalp_care_and_cleansing_products,
-        other_hair_and_scalp_care_and_cleansing_products: :hair_and_scalp_care_and_cleansing_products,
-        hair_colouring_products: :hair_and_scalp_products,
-        oxidative_hair_colour_products: :hair_colouring_products,
-        nonoxidative_hair_colour_products: :hair_colouring_products,
-        hair_bleaching_and_dye_remover_products: :hair_colouring_products,
-        other_hair_colouring_products: :hair_colouring_products,
-        hair_styling_products: :hair_and_scalp_products,
-        products_for_temporary_hair_styling: :hair_styling_products,
-        permanent_wave_products: :hair_styling_products,
-        hair_relaxer_straightener_products: :hair_styling_products,
-        other_hair_styling_products: :hair_styling_products,
-        other_hair_and_scalp_products: :hair_and_scalp_products,
-        hair_sun_protection_products: :other_hair_and_scalp_products,
-        other_hair_and_scalp_products_child: :other_hair_and_scalp_products,
-        nail_varnish_and_remover_products: :nail_and_cuticle_products,
-        nail_varnish_nail_makeup: :nail_varnish_and_remover_products,
-        nail_varnish_remover: :nail_varnish_and_remover_products,
-        nail_varnish_thinner: :nail_varnish_and_remover_products,
-        nail_bleach: :nail_varnish_and_remover_products,
-        other_nail_varnish_and_remover_products: :nail_varnish_and_remover_products,
-        nail_care_nail_hardener_products: :nail_and_cuticle_products,
-        nail_care_products: :nail_care_nail_hardener_products,
-        nail_hardener: :nail_care_nail_hardener_products,
-        other_nail_care_nail_hardener_products: :nail_care_nail_hardener_products,
-        nail_glue_remover_products: :nail_and_cuticle_products,
-        nail_glue_remover: :nail_glue_remover_products,
-        other_nail_and_cuticle_products: :nail_and_cuticle_products,
-        cuticle_remover_softener: :other_nail_and_cuticle_products,
-        nail_sculpting_products: :other_nail_and_cuticle_products,
-        other_nail_and_cuticle_products_child: :other_nail_and_cuticle_products,
-        tooth_care_products: :oral_hygiene_products,
-        toothpaste: :tooth_care_products,
-        tooth_cleansing_powder_salt: :tooth_care_products,
-        other_tooth_care_products: :tooth_care_products,
-        mouth_wash_breath_spray: :oral_hygiene_products,
-        mouth_wash: :mouth_wash_breath_spray,
-        breath_spray: :mouth_wash_breath_spray,
-        other_mouth_wash_breath_spray_products: :mouth_wash_breath_spray,
-        tooth_whiteners: :oral_hygiene_products,
-        tooth_whiteners_child: :tooth_whiteners,
-        other_oral_hygiene_products: :oral_hygiene_products,
-        other_oral_hygiene_products_child: :other_oral_hygiene_products
-    }.freeze
-
-    enum TRIGGER_RULES_QUESTION: {
-        please_specify_the_percentage_weight_of_ethanol: "please_specify_the_percentage_weight_of_ethanol",
-        please_specify_the_percentage_weight_of_isopropanol: "please_specify_the_percentage_weight_of_isopropanol",
-        please_specify_the_inci_name_and_concentration_of_the_antidandruff_agents_if_antidandruff_agents_are_not_present_in_the_cosmetic_product_then_not_applicable_must_be_checked: "please_specify_the_inci_name_and_concentration_of_the_antidandruff_agents_if_antidandruff_agents_are_not_present_in_the_cosmetic_product_then_not_applicable_must_be_checked",
-        please_specify_the_inci_name_and_concentration_of_the_antihair_loss_agents_if_antihair_loss_agents_are_not_present_in_the_cosmetic_product_then_not_applicable_must_be_checked: "please_specify_the_inci_name_and_concentration_of_the_antihair_loss_agents_if_antihair_loss_agents_are_not_present_in_the_cosmetic_product_then_not_applicable_must_be_checked",
-        please_specify_the_inci_name_and_concentration_of_the_antipigmenting_and_depigmenting_agents_if_antipigmenting_and_depigmenting_agents_are_not_present_in_the_cosmetic_product_then_not_applicable_must_be_checked: "please_specify_the_inci_name_and_concentration_of_the_antipigmenting_and_depigmenting_agents_if_antipigmenting_and_depigmenting_agents_are_not_present_in_the_cosmetic_product_then_not_applicable_must_be_checked",
-        please_specify_the_inci_name_and_concentration_of_chemical_exfoliating_agents_if_chemical_exfoliating_agents_are_not_present_in_the_cosmetic_product_then_not_applicable_must_be_checked: "please_specify_the_inci_name_and_concentration_of_chemical_exfoliating_agents_if_chemical_exfoliating_agents_are_not_present_in_the_cosmetic_product_then_not_applicable_must_be_checked",
-        please_specify_the_exact_content_of_vitamin_a_or_its_derivatives_for_the_whole_product_if_the_level_of_vitamin_a_or_any_of_its_derivatives_does_not_exceed_020_calculated_as_retinol_or_if_the_amount_does_not_exceed_009_grams_calculated_as_retinol_or_if_vitamin_a_or_any_of_its_derivatives_are_not_present_in_the_product_then_not_applicable_must_be_checked: "please_specify_the_exact_content_of_vitamin_a_or_its_derivatives_for_the_whole_product_if_the_level_of_vitamin_a_or_any_of_its_derivatives_does_not_exceed_020_calculated_as_retinol_or_if_the_amount_does_not_exceed_009_grams_calculated_as_retinol_or_if_vitamin_a_or_any_of_its_derivatives_are_not_present_in_the_product_then_not_applicable_must_be_checked",
-        please_specify_the_inci_name_and_the_concentration_of_xanthine_derivatives_eg_caffeine_theophylline_theobromine_plant_extracts_containing_xanthine_derivatives_eg_paulinia_cupana_guarana_extractspowders_if_xanthine_derivatives_are_not_present_or_present_below_05_in_the_cosmetic_product_then_not_applicable_must_be_checked: "please_specify_the_inci_name_and_the_concentration_of_xanthine_derivatives_eg_caffeine_theophylline_theobromine_plant_extracts_containing_xanthine_derivatives_eg_paulinia_cupana_guarana_extractspowders_if_xanthine_derivatives_are_not_present_or_present_below_05_in_the_cosmetic_product_then_not_applicable_must_be_checked",
-        please_specify_the_inci_name_and_concentration_of_the_cationic_surfactants_with_two_or_more_chain_lengths_below_c12_if_the_surfactant_is_used_for_non_preservative_purpose_if_cationic_surfactants_with_two_or_more_chain_lengths_below_c12_are_not_present_in_the_product_then_not_applicable_must_be_checked: "please_specify_the_inci_name_and_concentration_of_the_cationic_surfactants_with_two_or_more_chain_lengths_below_c12_if_the_surfactant_is_used_for_non_preservative_purpose_if_cationic_surfactants_with_two_or_more_chain_lengths_below_c12_are_not_present_in_the_product_then_not_applicable_must_be_checked",
-        please_specify_the_inci_name_and_concentration_of_each_propellant_if_propellants_are_not_present_in_the_product_then_not_applicable_must_be_checked: "please_specify_the_inci_name_and_concentration_of_each_propellant_if_propellants_are_not_present_in_the_product_then_not_applicable_must_be_checked",
-        please_specify_the_concentration_of_hydrogen_peroxide_if_hydrogen_peroxide_is_not_present_in_the_product_then_not_applicable_must_be_checked_: "please_specify_the_concentration_of_hydrogen_peroxide_if_hydrogen_peroxide_is_not_present_in_the_product_then_not_applicable_must_be_checked_",
-        please_specify_the_inci_name_and_the_concentration_of_the_compounds_that_release_hydrogen_peroxide_if_compounds_releasing_hydrogen_peroxide_are_not_present_in_the_product_then_not_applicable_must_be_checked: "please_specify_the_inci_name_and_the_concentration_of_the_compounds_that_release_hydrogen_peroxide_if_compounds_releasing_hydrogen_peroxide_are_not_present_in_the_product_then_not_applicable_must_be_checked",
-        please_specify_the_inci_name_and_concentration_of_each_reducing_agent_if_reducing_agents_are_not_present_in_the_product_then_not_applicable_must_be_checked: "please_specify_the_inci_name_and_concentration_of_each_reducing_agent_if_reducing_agents_are_not_present_in_the_product_then_not_applicable_must_be_checked",
-        please_specify_the_inci_name_and_concentration_of_each_persulfate_if_persulfates_are_not_present_in_the_product_then_not_applicable_must_be_checked: "please_specify_the_inci_name_and_concentration_of_each_persulfate_if_persulfates_are_not_present_in_the_product_then_not_applicable_must_be_checked",
-        please_specify_the_inci_name_and_concentration_of_each_straightening_agent_if_straightening_agents_are_not_present_in_the_product_then_not_applicable_must_be_checked: "please_specify_the_inci_name_and_concentration_of_each_straightening_agent_if_straightening_agents_are_not_present_in_the_product_then_not_applicable_must_be_checked",
-        please_indicate_the_total_concentration_of_inorganic_sodium_salts_if_inorganic_sodium_salts_are_not_present_in_the_product_then_not_applicable_must_be_checked: "please_indicate_the_total_concentration_of_inorganic_sodium_salts_if_inorganic_sodium_salts_are_not_present_in_the_product_then_not_applicable_must_be_checked",
-        please_indicate_the_concentration_of_fluoride_compounds_calculated_as_fluorine_if_fluoride_compounds_are_not_present_in_the_product_then_not_applicable_must_be_checked: "please_indicate_the_concentration_of_fluoride_compounds_calculated_as_fluorine_if_fluoride_compounds_are_not_present_in_the_product_then_not_applicable_must_be_checked",
-        is_the_ph_of_the_component_lower_than_3_or_higher_than_10: "is_the_ph_of_the_component_lower_than_3_or_higher_than_10",
-        please_indicate_the_total_level_of_essential_oils_camphor_menthol_or_eucalyptol_if_essential_oils_camphor_menthol_or_eucalyptol_are_not_present_in_the_product_or_if_the_level_of_essential_oils_camphor_menthol_or_eucalyptol_does_not_exceed_05_then_not_applicable_must_be_checked: "please_indicate_the_total_level_of_essential_oils_camphor_menthol_or_eucalyptol_if_essential_oils_camphor_menthol_or_eucalyptol_are_not_present_in_the_product_or_if_the_level_of_essential_oils_camphor_menthol_or_eucalyptol_does_not_exceed_05_then_not_applicable_must_be_checked",
-        please_indicate_the_name_and_the_quantity_of_each_essential_oil_camphor_menthol_or_eucalyptol_if_no_individual_essential_oil_camphor_menthol_or_eucalyptol_are_present_with_a_level_higher_than_05_015_in_case_of_camphor_then_not_applicable_must_be_checked: "please_indicate_the_name_and_the_quantity_of_each_essential_oil_camphor_menthol_or_eucalyptol_if_no_individual_essential_oil_camphor_menthol_or_eucalyptol_are_present_with_a_level_higher_than_05_015_in_case_of_camphor_then_not_applicable_must_be_checked",
-        please_indicate_the_total_concentration_of_inorganic_sodium_salts: "please_indicate_the_total_concentration_of_inorganic_sodium_salts",
-        please_indicate_the_ph_of_the_hair_dye_component: "please_indicate_the_ph_of_the_hair_dye_component",
-        please_indicate_the_ph_of_the_mixed_hair_dye_product: "please_indicate_the_ph_of_the_mixed_hair_dye_product",
-        please_indicate_the_ph: "please_indicate_the_ph",
-        please_indicate_the_ph_of_the_mixed_product_: "please_indicate_the_ph_of_the_mixed_product_",
-        do_the_components_of_the_product_need_to_be_mixed: "do_the_components_of_the_product_need_to_be_mixed",
-        please_indicate_the_inci_name_and_concentration_of_each_alkaline_agent_including_ammonium_hydroxide_liberators: "please_indicate_the_inci_name_and_concentration_of_each_alkaline_agent_including_ammonium_hydroxide_liberators"
-    }
-
-    enum TRIGGER_RULES_QUESTION_ELEMENT: {
-        ethanol: "ethanol",
-        propanol: "propanol",
-        inciname: "inciname",
-        incivalue: "incivalue",
-        value: "value",
-        ph: "ph",
-        concentration: "concentration",
-        minrangevalue: "minrangevalue",
-        maxrangevalue: "maxrangevalue"
-    }
-  end
-end
-# rubocop:enable Metrics/BlockLength
