+++ cosmetics-web/db/migrate/20190220140100_add_formulation_to_components.rb
@@ -14,7 +14,7 @@ class AddFormulationToComponents < ActiveRecord::Migration[5.2]
 
     create_table :range_formulas do |t|
       t.string :inci_name
-      t.integer :range
+      t.string :range
 
       t.timestamps
     end
