+++ cosmetics-web/lib/tasks/cpnp_import.rake
@@ -1,77 +1,82 @@
 namespace :cpnp_import do
   desc "Print static product data from CPNP"
   task :categories do
-    print_three_mapping_data_structures('category', 'categories', 'category')
+    print_three_mapping_data_structures("category", "categories", "category")
   end
 
   task :categories_parent do
     parent_hash_command = start_hash_command("PARENT_OF_CATEGORY")
 
-    xml_file_content = File.read(Rails.root.join('app', 'assets', 'files', 'cpnp', 'categories.xml'))
+    xml_file_content = File.read(Rails.root.join("app", "assets", "files", "cpnp", "categories.xml"))
     xml_doc = Nokogiri::XML(xml_file_content)
 
-    underlined_names_set = Set.new
+    category_key_names_set = Set.new
     xml_doc.xpath("//category[lngId='EN']").each do |english_category_node|
-      english_category_name = english_category_node.xpath('name').first&.text
-      underlined_name = get_underline_separated_name(english_category_name)
-      while underlined_names_set.add?(underlined_name).blank?
-        underlined_name = underlined_name + '_child'
+      english_category_name = english_category_node.xpath("name").first&.text
+      category_key_name = get_category_key_name(english_category_name)
+      while category_key_names_set.add?(category_key_name).blank?
+        category_key_name = category_key_name + "_child"
       end
 
-      parent_id = english_category_node.xpath('parentID').first&.text
+      parent_id = english_category_node.xpath("parentID").first&.text
       if parent_id.present?
         parent_english_category_name = xml_doc.xpath("//category[id=#{parent_id} and lngId='EN']/name").first&.text
-        parent_underlined_name = get_underline_separated_name(parent_english_category_name)
-        parent_hash_command += string_label_pair(underlined_name, parent_underlined_name)
+        parent_category_key_name = get_category_key_name(parent_english_category_name)
+        parent_hash_command += string_label_pair(category_key_name, parent_category_key_name)
       end
     end
     puts(end_command(parent_hash_command))
   end
 
   task :frame_formulations do
-    print_three_mapping_data_structures('frame_formulation', 'frame_formulation', 'frameFormulation')
+    print_three_mapping_data_structures("frame_formulation", "frameFormulation", "frameFormulation")
   end
 
   task :trigger_rules_questions do
-    print_three_mapping_data_structures('trigger_rules_question', 'questions', 'question', 'id', 'description', false)
+    print_three_mapping_data_structures("trigger_rules_question", "questions", "question", "id", "description", false)
   end
 
   task :trigger_rules_question_elements do
-    print_three_mapping_data_structures('trigger_rules_question_element', 'questions', 'element', 'elemID', 'elemName', false, false)
+    print_three_mapping_data_structures("trigger_rules_question_element", "questions", "element", "elemID", "elemName", false, false)
+  end
+
+  task :units do
+    print_three_mapping_data_structures("unit", "units", "unit", "id", "name", false, false, true)
   end
 end
 
-def print_three_mapping_data_structures(variable_name, file_name, tag_name, id_tag_name = 'id', description_tag_name = 'name', enable_lang = true, enable_duplication = true)
+def print_three_mapping_data_structures(variable_name, file_name, main_tag, id_tag_name = "id", description_tag_name = "name",
+                                        enable_lang = true, enable_duplication = true, enable_math_translation = false)
   variable_name.upcase!
   enum_command = start_enum_command(variable_name)
   mapping_hash_command = start_hash_command("#{variable_name}_ID")
   view_hash_command = start_hash_command("#{variable_name}_NAME")
 
-  xml_file_content = File.read(Rails.root.join('app', 'assets', 'files', 'cpnp', "#{file_name}.xml"))
+  xml_file_content = File.read(Rails.root.join("app", "assets", "files", "cpnp", "#{file_name}.xml"))
   xml_doc = Nokogiri::XML(xml_file_content)
 
-  underlined_names_set = Set.new
+  category_key_names_set = Set.new
 
-  query = "//#{tag_name}" + (enable_lang ? "[lngId='EN']" : "")
+  query = "//#{main_tag}" + (enable_lang ? "[lngId='EN']" : "")
   xml_doc.xpath(query).each do |english_category_node|
     english_category_name = english_category_node.xpath(description_tag_name).first&.text
     id = english_category_node.xpath(id_tag_name).first&.text
-    underlined_name = get_underline_separated_name(english_category_name)
+    category_key_name = get_category_key_name(english_category_name, enable_math_translation)
 
-    underlined_name_exist_in_set = underlined_names_set.add?(underlined_name).blank?
-    if underlined_name_exist_in_set
+    category_key_name_exist_in_set = category_key_names_set.add?(category_key_name).blank?
+    if category_key_name_exist_in_set
       if enable_duplication
-        while underlined_names_set.add?(underlined_name).blank?
-          underlined_name = underlined_name + '_child'
+        while category_key_names_set.add?(category_key_name).blank?
+          category_key_name = category_key_name + "_child"
         end
       end
     end
-    if !underlined_name_exist_in_set || enable_duplication
-      enum_command += string_string_pair(underlined_name, underlined_name)
-      view_hash_command += string_string_pair(underlined_name, english_category_name.delete('"'))
+    if !category_key_name_exist_in_set || enable_duplication
+      enum_command += string_string_pair(category_key_name, category_key_name)
+      view_hash_command += string_string_pair(category_key_name, english_category_name.delete('"'))
     end
 
-    mapping_hash_command += int_string_pair(id, underlined_name)
+    mapping_hash_command += int_string_pair(id, category_key_name)
   end
 
   puts(end_command(enum_command))
@@ -96,13 +101,19 @@ def string_string_pair(str1, str2)
 end
 
 def string_label_pair(str1, label2)
-  "\t" + str1 + ': :' + label2 + ",\n"
+  "\t" + str1 + ": :" + label2 + ",\n"
 end
 
 def int_string_pair(int1, str2)
-  "\t" + int1 + ' => :' + str2 + ",\n"
+  "\t" + int1 + " => :" + str2 + ",\n"
 end
 
-def get_underline_separated_name(name)
-  name.downcase.gsub(/[^a-zA-Z0-9\s]+/, '').gsub(/\s+/, '_')
+def get_category_key_name(name, enable_math_translation = false)
+  result = name
+  if enable_math_translation
+    result = name.gsub(/[≤<]+/, "less than")
+                 .gsub(/[≥>]+/, "greater than")
+                 .gsub(/[%]+/, " percent")
+  end
+  result.downcase.gsub(/[^a-zA-Z0-9\s]+/, "").gsub(/\s+/, "_")
 end
